% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chisqtest.R
\name{chisqtest}
\alias{chisqtest}
\title{chisqtest function}
\usage{
chisqtest(filter.data, tUTR.gene, bigene, bilabel)
}
\arguments{
\item{filter.data}{a PA dataframe}

\item{tUTR.gene}{all gene in dataframe}

\item{bigene}{gene which belong to Bimodal}

\item{bilabel}{label which belong to Bimodal}
}
\description{
This function allows you to test Bimodal.
}
\keyword{chisqtest}
