% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut2.R
\name{cut2}
\alias{cut2}
\title{cut2 function}
\usage{
cut2(
  x,
  breaks,
  labels = NULL,
  include.lowest = FALSE,
  right = TRUE,
  dig.lab = 3L,
  ordered_result = FALSE,
  fixed_max,
  ...
)
}
\arguments{
\item{x}{a matrix}

\item{breaks}{set space}

\item{labels}{label of data or the color of cells}

\item{include.lowest}{default FALSE}

\item{right}{default TRUE}

\item{dig.lab}{the maxinum of sapce}

\item{ordered_result}{default FALSE}

\item{fixed_max}{the maxinum of data}
}
\description{
This function allows you to exploring the connection between data.
}
\keyword{cut2}
