% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exMajorPA.R
\name{exMajorPA}
\alias{exMajorPA}
\title{exMajorPA function}
\usage{
exMajorPA(scAPAtrapds, cell_type = "SC")
}
\arguments{
\item{scAPAtrapds}{a PACdataset}

\item{cell_type}{select the cell type}
}
\description{
This function allows you to calculate the ratio of PAs and select the maximum and minimum values.
}
\keyword{exMajorPA}
