% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrPairPA.R
\name{extrPairPA}
\alias{extrPairPA}
\title{extrPairPA function}
\usage{
extrPairPA(exc1.g1, all_gene, all_anno, gn = 4, cn = 10)
}
\arguments{
\item{exc1.g1}{counts from a PACdataset(APA expression in the same cell type)}

\item{all_gene}{gene from anno in PACdataset}

\item{all_anno}{anno from a PACdataset}

\item{gn}{proportion of filtered cells}

\item{cn}{proportion of filtered genes}
}
\description{
This function allows you to filter PAs and cells, and calculate the PUI of genes.
}
\keyword{extrPairPA}
