% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filftr.R
\name{filftr}
\alias{filftr}
\title{filftr function}
\usage{
filftr(scAPAtrapds)
}
\arguments{
\item{scAPAtrapds}{a PACdataset}
}
\description{
This function allows you to filter out PAs in the intergenic region.
}
\keyword{filftr}
