% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filgene.R
\name{filgene}
\alias{filgene}
\title{filgene function}
\usage{
filgene(scAPAtrapds)
}
\arguments{
\item{scAPAtrapds}{a PACdataset}
}
\description{
This function allows you to remove genes named character(0).
}
\keyword{filgene}
