% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filPAs.R
\name{filPAs}
\alias{filPAs}
\title{filPAs function}
\usage{
filPAs(PAcounts, scAPAtrapds.tmp, gn = 4, cn = 10)
}
\arguments{
\item{PAcounts}{the expression of PA (>=2)}

\item{scAPAtrapds.tmp}{a PACdataset}

\item{gn}{proportion of filtered cells}

\item{cn}{proportion of filtered genes or PAs}
}
\description{
This function allows you to filter PAs and cells.
}
\keyword{filPAs}
