% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filPAs2.R
\name{filPAs2}
\alias{filPAs2}
\title{filPAs2 function}
\usage{
filPAs2(PAcounts, gene, ftr, gn = 4, cn = 10)
}
\arguments{
\item{PAcounts}{the expression of PA (>=2)}

\item{gene}{gene corresponding to PAcounts}

\item{ftr}{the location of PA corresponding to PAcounts}

\item{gn}{proportion of filtered cells}

\item{cn}{proportion of filtered genes or PAs}
}
\description{
This function allows you to filter PAs and cells.
}
\keyword{filPAs2}
