% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMMod.R
\name{getMMod}
\alias{getMMod}
\title{getMMod function}
\usage{
getMMod(APAdata, type)
}
\arguments{
\item{type}{select "PAmax" or "PAmin"}

\item{APAData}{a PACdataset}
}
\description{
This function allows you to identify APA usage modalities.
The model used in the function is Gaussion Mixture model.
}
\keyword{getMMod}
