% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMod.R
\name{getMod}
\alias{getMod}
\title{getMod function}
\usage{
getMod(PUIData)
}
\arguments{
\item{PUIData}{a matrix which is the result of extrPairPA or exnon3UTRPA}
}
\description{
This function allows you to identify APA usage modalities.
The model used in the function is Gaussion Mixture model.
}
\keyword{getMod}
