% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KStest.R
\name{KStest}
\alias{KStest}
\title{KStest function}
\usage{
KStest(PUIdata, tUTR.gene, ftr, bigene, bilabel)
}
\arguments{
\item{PUIdata}{a PUI dataframe}

\item{tUTR.gene}{all gene in dataframe}

\item{ftr}{all ftr in dataframe}

\item{bigene}{gene which belong to Bimodal}

\item{bilabel}{label which belong to Bimodal}
}
\description{
This function allows you to test Bimodal.
}
\keyword{KStest}
