% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAMIMod.R
\name{MAMIMod}
\alias{MAMIMod}
\title{MAMIMod function}
\usage{
MAMIMod(APAdata, celltype, type)
}
\arguments{
\item{APAdata}{a PACdataset}

\item{celltype}{the index of one same celltype}

\item{type}{select "MajorPA" or "MinorPA"}
}
\description{
This function allows you to analyze APA's preferences.
The model used in the function is Gaussion Mixture model.
}
\keyword{MAMIMod}
