% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotColorKey2.R
\name{plotColorKey2}
\alias{plotColorKey2}
\title{plotColorKey2 function}
\usage{
plotColorKey2(colors, min_color, max_color)
}
\arguments{
\item{colors}{select color}

\item{min_color}{the minimum value of PA expression}

\item{max_color}{the maximum value of PA expression}
}
\description{
This function allows you to add a label about expressions of PAs.
}
\keyword{plotColorKey2}
