% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotColorLabel.R
\name{plotColorLabel}
\alias{plotColorLabel}
\title{plotColorLabel function}
\usage{
plotColorLabel(color_labels)
}
\arguments{
\item{color_labels}{display gene name or PA information}
}
\description{
This function allows you to display gene name or PA information.
}
\keyword{plotColorLabel}
