% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotComHeatmap.R
\name{plotComHeatmap}
\alias{plotComHeatmap}
\title{plotComHeatmap function}
\usage{
plotComHeatmap(mat, breaks, pal, fixed_max)
}
\arguments{
\item{mat}{a matrix}

\item{breaks}{set spacing}

\item{pal}{set color}

\item{fixed_max}{set maxinum value}
}
\description{
This function allows you to draw the PAs expression heat map of each component.
}
\keyword{plotComHeatmap}
