% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotGenePACount3.R
\name{plotGenePACount3}
\alias{plotGenePACount3}
\title{plotGenePACount3 function}
\usage{
plotGenePACount3(org.Mm.eg.db, id, tUTR.pair.cd.tmp, tUTR.gene.fil, label)
}
\arguments{
\item{org.Mm.eg.db}{select an annotation information for a species}

\item{id}{index belonging to Bimodal}

\item{tUTR.pair.cd.tmp}{PAs expression(filter.data-output of extrPairPA)}

\item{tUTR.gene.fil}{gene sample}

\item{label}{label of cell components}
}
\description{
This function allows you to draw PAs expression heat map and expression heat map of each component.
}
\keyword{plotGenePACount3}
