% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHeatmap.R
\name{plotHeatmap}
\alias{plotHeatmap}
\title{plotHeatmap function}
\usage{
plotHeatmap(mat, breaks, pal, fixed_max)
}
\arguments{
\item{mat}{a matrix}

\item{breaks}{set spacing}

\item{pal}{set color}

\item{fixed_max}{set maximum}
}
\description{
This function allows you to draw PAs expression heat map.
}
\keyword{plotHeatmap}
