% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotYvalue.R
\name{plotYvalue}
\alias{plotYvalue}
\title{plotYvalue function}
\usage{
plotYvalue(nr, max_color, min_color)
}
\arguments{
\item{nr}{Number of components}

\item{max_color}{Maximum value of each component}

\item{min_color}{Minimum value of each component}
}
\description{
This function allows you to calculate and display the maximum and minimum values of expression.
}
\keyword{plotYvalue}
