% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDiffusionMap.R
\name{runDiffusionMap}
\alias{runDiffusionMap}
\title{runDiffusionMap function}
\usage{
runDiffusionMap(mat)
}
\arguments{
\item{mat}{a Matrix}
}
\description{
This function allows you to sort by expression
}
\keyword{runDiffusionMap}
