% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UTRmod.R
\name{UTRmod}
\alias{UTRmod}
\title{UTRmod function}
\usage{
UTRmod(APAdata, celltype)
}
\arguments{
\item{APAdata}{a PACdataset}

\item{celltype}{the index of one same celltype}
}
\description{
This function allows you to identifying modalities of APA usage in 3UTR.
The model used in the function is Gaussion Mixture model.
}
\keyword{UTRmod}
